feat: add main page img, player stat querymethod
