+++ src/main/resources/templates/basketball-league-ranking-team.html
@@ -37,7 +37,7 @@
     </nav>
 
     <!-- Image Header -->
-    <img class="img-fluid rounded mb-4" src="http://placehold.it/1200x300" alt="">
+    <img class="img-fluid rounded mb-4" src="/img/basketball_stat_img2.jpg" alt="">
 
     <!-- Marketing Icons Section -->
     <div>
