+++ src/main/resources/templates/basketball-league-tournament.html
@@ -9,7 +9,7 @@
 
     <!-- Page Heading/Breadcrumbs -->
     <h1 class="mt-4 mb-3">농구
-        <small>경기일정 및 결과</small>
+        <small>대회소개</small>
     </h1>
 
     <ol class="breadcrumb">
@@ -19,7 +19,7 @@
         <li class="breadcrumb-item">
             <a href="basketball">Basketball</a>
         </li>
-        <li class="breadcrumb-item active">경기일정 및 결과</li>
+        <li class="breadcrumb-item active">대회소개</li>
     </ol>
 
     <nav class="navbar-brand">
