+++ src/main/resources/templates/index.html
@@ -100,16 +100,24 @@
             <li data-target="#carouselExampleIndicators" data-slide-to="0" class="active"></li>
             <li data-target="#carouselExampleIndicators" data-slide-to="1"></li>
             <li data-target="#carouselExampleIndicators" data-slide-to="2"></li>
+            <li data-target="#carouselExampleIndicators" data-slide-to="3"></li>
+            <li data-target="#carouselExampleIndicators" data-slide-to="3"></li>
         </ol>
         <div class="carousel-inner" role="listbox">
             <!-- Slide One - Set the background image for this slide in the line below -->
-            <div class="carousel-item active" style="background-image: url('/img/index_img_01.jpeg');">
+            <div class="carousel-item active" style="background-image: url('/img/main1.jpg');">
             </div>
             <!-- Slide Two - Set the background image for this slide in the line below -->
-            <div class="carousel-item" style="background-image: url('/img/index_img_02.jpeg');">
+            <div class="carousel-item" style="background-image: url('/img/main2.jpeg');">
             </div>
             <!-- Slide Three - Set the background image for this slide in the line below -->
-            <div class="carousel-item" style="background-image: url('/img/index_img_03.jpeg');">
+            <div class="carousel-item" style="background-image: url('/img/main3.jpg');">
+            </div>
+            <!-- Slide Four - Set the background image for this slide in the line below -->
+            <div class="carousel-item" style="background-image: url('/img/main4.jpeg');">
+            </div>
+            <!-- Slide Five - Set the background image for this slide in the line below -->
+            <div class="carousel-item" style="background-image: url('/img/main5.jpeg');">
             </div>
         </div>
         <a class="carousel-control-prev" href="#carouselExampleIndicators" role="button" data-slide="prev">
